Split BankCardNumberUtils into BankCardNumberUtils and StringUtils. Fix allowed card number whitespaces character set.
